backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
numpy==1.23.5
tensorflow==2.12.0
pillow==10.0.1
opencv-python-headless==4.8.1.78
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
easyocr==1.7.1
scikit-learn==1.3.2
scipy==1.10.1
transformers==4.41.2
sentence-transformers==2.2.2
requests==2.31.0
tqdm==4.66.1
timm==0.9.12